deps: Update to Next 16 and TS 6
