+++ package.json
@@ -12,10 +12,10 @@
   "dependencies": {
     "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^3.0.0",
-    "@next/mdx": "^15.0.0",
+    "@next/mdx": "^16.2.4",
     "@tailwindcss/postcss": "^4.0.0",
     "mdast-util-find-and-replace": "^3.0.0",
-    "next": "^15.5.15",
+    "next": "^16.2.4",
     "postcss": "^8.4.18",
     "react": "19.2.5",
     "react-dom": "19.2.5",
@@ -31,7 +31,7 @@
     "@types/node": "24.12.2",
     "@types/react": "19.2.14",
     "prettier": "3.8.3",
-    "typescript": "5.9.3",
+    "typescript": "^6.0.1",
     "vercel": "^52.0.0"
   },
   "packageManager": "yarn@4.14.1",
